marlin: bf16 scales by default (match model dtype), drop fp16 conversions
marlin requires scale dtype == activation dtype (bf16-a + fp16-scales = garbage,
rel-err 1.0; bf16-a + bf16-scales = correct, 0.0028). quantize_to_marlin_int4 now
defaults scale_dtype=bf16 (the model dtype), so the gate_up decode integration
passes bf16 hidden directly with NO per-layer bf16<->fp16 conversion. Test + the
dequant-ref check updated to bf16. Sets up the full all-4-GEMM int4 decode (next).

+++ tests/test_marlin_int4.py
@@ -33,8 +33,8 @@ def test_marlin_int4_mm_matches_dequant_reference():
     # gate_up and down shapes (N % 64 == 0, K % 16 == 0)
     for N, K in [(7168, 8192), (8192, 3584)]:
         assert marlin.marlin_supports_shape(N, K)
-        w = torch.randn(K, N, device=dev, dtype=torch.float16) * 0.02
-        a = torch.randn(48, K, device=dev, dtype=torch.float16)
+        w = torch.randn(K, N, device=dev, dtype=torch.bfloat16) * 0.02
+        a = torch.randn(48, K, device=dev, dtype=torch.bfloat16)
         q, s = marlin.quantize_to_marlin_int4(w, 128)
         ws = marlin.make_workspace(N, dev)
         out = marlin.marlin_int4_mm(a, q, s, ws, N, K)
